fix safe warning

+++ modules/skshaper/src/SkShaper_harfbuzz.cpp
@@ -778,7 +778,7 @@ struct ShapedRunGlyphIterator {
         if (fGlyphIndex == runs[fRunIndex].fNumGlyphs) {
             fGlyphIndex = 0;
             ++fRunIndex;
-            if (fRunIndex >= runs.size()) {
+            if (static_cast<size_t>(fRunIndex) >= runs.size()) {
                 return nullptr;
             }
         }
@@ -787,7 +787,7 @@ struct ShapedRunGlyphIterator {
 
     ShapedGlyph* current() {
         const SkTArray<ShapedRun>& runs = *fRuns;
-        if (fRunIndex >= runs.size()) {
+        if (static_cast<size_t>(fRunIndex) >= runs.size()) {
             return nullptr;
         }
         return &runs[fRunIndex].fGlyphs[fGlyphIndex];
